% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_warwick.R
\name{theme_warwick}
\alias{theme_warwick}
\title{ggplot2 theme consistent with Warwick Brand}
\usage{
theme_warwick(base_size = 11)
}
\arguments{
\item{base_size}{numeric, base size for font}
}
\description{
ggplot2 theme consistent with Warwick Brand
}
\examples{
\dontrun{
ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(cty, hwy)) +
   ggplot2::geom_point() +
   theme_warwick()
}
}
