% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{warwick_palettes}
\alias{warwick_palettes}
\alias{warwick_palettes_discrete}
\alias{warwick_palettes_sequential}
\alias{warwick_palettes_divergent}
\title{University of Warwick brand palettes}
\format{
\subsection{\code{warwick_palettes}}{

An object of class \code{palettes_palette} with 6
colour palettes.
}

\subsection{\code{warwick_palettes_discrete}}{

An object of class \code{palettes_palette} with
1 discrete colour palettes.
}

\subsection{\code{warwick_palettes_sequential}}{

An object of class \code{palettes_palette} with
3 sequential colour palettes.
}

\subsection{\code{warwick_palettes_divergent}}{

An object of class \code{palettes_palette} with
2 sequential colour palettes.
}
}
\source{
\url{https://github.com/Warwick-Stats-Resources/warwickplots}
}
\usage{
warwick_palettes

warwick_palettes_discrete

warwick_palettes_sequential

warwick_palettes_divergent
}
\description{
Palettes consistent with The University of Warwick's brand.
}
\details{
For more information on the University of Warwick's brand colours, see \url{https://warwick.ac.uk/about/brand/brand-guidelines/colours/}.

The primary palette is a combination of the primary brand colour (aubergine) and the five bright accent colours.
These are the colours that the brand team uses for the graphs in their PowerPoint template.

The three sequential palettes run between the bright colour and its 30\% tint.

Since the primary Warwick brand aubergine is much darker than the bright accent colours, it was too strong as a starting point for the aubergine side of the divergent palettes. To achieve a better balance, \code{aubergine_ruby} starts with an 80\% tint of the primary aubergine, and \code{aubergine_teal} starts with a 70\% tint.
}
\examples{
# Get all palettes by name
names(warwick_palettes)

# Plot a specific palette
plot(warwick_palettes$primary)

# Subset a palette
warwick_palettes$primary[c(1, 3, 5)]

#' # Plot all palettes
\dontrun{
plot(warwick_palettes)
}
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/EllaKaye}{Ella Kaye}
}
\keyword{datasets}
