% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warwickplots-package.R
\docType{package}
\name{warwickplots-package}
\alias{warwickplots}
\alias{warwickplots-package}
\title{warwickplots: Palettes and themes aligned with The University of Warwick's brand}
\description{
Colour palettes and 'ggplot2' themes that are consistent with The University of Warwick's ID7 branding. Built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, colour mixing arithmetic, and colour interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Warwick-Stats-Resources/warwickplots}
  \item Report bugs at \url{https://github.com/Warwick-Stats-Resources/warwickplots/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.m.kaye@warwick.ac.uk} (\href{https://orcid.org/0000-0002-7300-3718}{ORCID})

}
\keyword{internal}
