% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warwickplots-package.R
\docType{package}
\name{warwickplots-package}
\alias{warwickplots}
\alias{warwickplots-package}
\title{warwickplots: Palettes and themes consistent with The University of Warwick's brand}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Colour palettes and a 'ggplot2' theme that are consistent with The University of Warwick' branding. Built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, colour mixing arithmetic, and colour interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Warwick-Stats-Resources/warwickplots}
  \item \url{https://warwick-stats-resources.github.io/warwickplots/}
  \item Report bugs at \url{https://github.com/Warwick-Stats-Resources/warwickplots/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.m.kaye@warwick.ac.uk} (\href{https://orcid.org/0000-0002-7300-3718}{ORCID})

}
\keyword{internal}
